+++ patches/src/main/kotlin/dev/jkcarino/revanced/patches/google/gboard/detection/signature/BypassSignaturePatch.kt
@@ -1,20 +1,17 @@
 package dev.jkcarino.revanced.patches.google.gboard.detection.signature
 
-import app.revanced.patcher.extensions.InstructionExtensions.addInstructions
 import app.revanced.patcher.patch.bytecodePatch
+import dev.jkcarino.revanced.patches.google.gboard.fixes.applyWorkaroundPatch
+import dev.jkcarino.revanced.util.returnEarly
 
 val bypassSignaturePatch = bytecodePatch(
     description = "Bypasses the signature verification checks for Gboard.",
 ) {
+    dependsOn(applyWorkaroundPatch)
+
     compatibleWith("com.google.android.inputmethod.latin")
 
     execute {
-        checkSignatureFingerprint.method.addInstructions(
-            index = 0,
-            smaliInstructions = """
-                const/4 v0, 0x1
-                return v0
-            """
-        )
+        checkSignatureFingerprint.method.returnEarly(true)
     }
 }